Bomber Enemy AI almost finished
The bomber AI enemy is almost finished, it now moves within range, faces the player automatically, and fires on the player on a specified time interval. The bomber projectiles have also been implemented and work as desired. The projectiles can be dodged as they just fly in a straight line.